# .history/app/Http/Controllers/Api/V1/ProductApiController_20221216105520.php
<?php

namespace App\Http\Controllers\Api\V1;

use App\Models\Product;
use Illuminate\Http\Request;
use App\Http\Controllers\Controller;
use App\Http\Resources\ProductResource;
use App\Repositories\ProductRepository;
use Illuminate\Support\Facades\Validator;

class ProductApiController extends Controller
{

    /**
     * @var ProductRepository
     */
    protected $productRepository;

    public function __construct(ProductRepository $productRepository)
    {
        $this->productRepository = $productRepository;
    }

    /**
     * @OA\Get(
     *   path="/api/v1/products",
     *   operationId="getAllProducts",
     *   summary="Fetch All Products",
     *   tags={"Products"},
     *   description="Get list of products",
     *   security={ {"BearerAuth": {} }},
     *   @OA\Response(
     *      response=200, 
     *      description="successful operation",
     *      @OA\MediaType(
     *         mediaType="application/json",
     *      ),
     *   ),
     *   @OA\Response(
     *      response=401,
     *      description="Unauthorized",
     *   ),
     *   @OA\Response(
     *       response=403,
     *        description="Forbidden"
     *    )
     * ),
     * @author Boosuro Stephen <boosurostephen@yahoo.com>
     * 
     * Display a listing of the Products.
     * 
     *
     * @param Request $request
     * @return \Illuminate\Http\JsonResponse
     */
    public function index()
    {
        $products = $this->productRepository->all();

        return $this->sendResponse($products, 'Products retrieved successfully', 200);
    }


    /**
     * @OA\Post(
     *   path="/api/v1/products",
     *   operationId="storeProduct",
     *   summary="Store a product",
     *   tags={"Products"},
     *   description="Store Product",
     *   security={ {"BearerAuth": {} }},
     *   @OA\RequestBody(
     *       required=true,
     *       @OA\MediaType(
     *           mediaType="multipart/form-data",
     *           @OA\Schema(
     *               type="object",
     *               @OA\Property(
     *                   property="name",
     *                   description="Product Name",
     *                   type="string",
     *                   example="Demo Iphone 14 Pro Max"
     *               ),
     *               @OA\Property(
     *                   property="price",
     *                   description="Demo Product Price",
     *                   type="string",
     *                   example="1111.99"
     *               ),
     *              @OA\Property(
     *                   property="quantity",
     *                   description="Product Quantity",
     *                   type="integer",
     *                   example="30"
     *               ),
     *              @OA\Property(
     *                   property="product_variants",
     *                   description="Product Variants",
     *                   type="string",
     *                   example="separate ids by comma eg. 1,2,3 etc",
     *               ),
     *              @OA\Property(
     *                  description="Product Image",
     *                  property="image",
     *                  type="file",
     *              ),
     *           )
     *       )
     *   ),   
     *  @OA\Response(
     *      response=201, 
     *      description="product created successfully",
     *      @OA\MediaType(
     *         mediaType="application/json",
     *      ),
     *   ),
     *   @OA\Response(
     *      response=401,
     *      description="Unauthorized",
     *   ),
     *   @OA\Response(
     *       response=403,
     *        description="Forbidden"
     *    )
     * ),
     * 
     * @author Boosuro Stephen <boosurostephen@yahoo.com>
     * 
     * Store a newly created resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return \Illuminate\Http\Response
     */
    public function store(Request $request)
    {

        // Validate Request
        $validator = Validator::make($request->only('name', 'price', 'quantity', 'product_variants', 'image'), [
            'name' => 'required|min:3',
            'price' => 'required|numeric|gt:0',
            'quantity' => 'required|numeric|min:0',
            'product_variants' => 'required|min:1',
            'image' => 'nullable|image|mimes:jpg,png,jpeg|max:2048'
        ]);

        // Check if validation failed
        if ($validator->fails()) {
            return $this->sendError($validator->errors(), 'Incomplete Data Provided', 422);
        }

        // Create Product
        $product = $this->productRepository->create(array_merge(
            $validator->validated(),
            ['user_id' => auth()->id()]
        ));

        $product_variants = explode(',', $validator->validated()['product_variants']);

        // Check if Image is part of request
        if ($request->hasFile('image')) {
            $product->addMediaFromRequest('image')->toMediaCollection('images');
        }

        // Fill Pivot table with data
        $product->variants()->attach($product_variants);

        return $this->sendResponse(new ProductResource($product), 'Product Created Successfully', 201);
    }

    /**
     * @OA\Get(
     *   path="/api/v1/products/{id}",
     *   operationId="getProduct",
     *   summary="Fetch a Product",
     *   tags={"Products"},
     *   description="Fetch a product",
     *   security={ {"BearerAuth": {} }},
     *   @OA\Parameter(
     *       name="id",
     *       description="Product id",
     *       required=true,
     *       in="path",
     *       @OA\Schema(
     *          type="integer"
     *       )
     *   ),
     *  @OA\Response(
     *      response=201, 
     *      description="product retrieved successfully",
     *      @OA\MediaType(
     *         mediaType="application/json",
     *      ),
     *   ),
     *   @OA\Response(
     *      response=401,
     *      description="Unauthorized",
     *   ),
     *   @OA\Response(
     *       response=403,
     *        description="Forbidden"
     *    )
     * )
     * @author Boosuro Stephen <boosurostephen@yahoo.com>
     * 
     * Display the specified resource.
     *
     * @param  \App\Models\Product  $product
     * @return \Illuminate\Http\Response
     */
    public function show(Product $product)
    {
        return $this->sendResponse(new ProductResource($product), 'Product Retrieved Successfully', 201);
    }

    /**
     * @OA\Put(
     *   path="/api/v1/products/{id}",
     *   operationId="updateProduct",
     *   summary="Update a product",
     *   tags={"Products"},
     *   description="Update Product",
     *   security={ {"BearerAuth": {} }},
     *   @OA\Parameter(
     *       name="id",
     *       description="Product id",
     *       required=true,
     *       in="path",
     *       @OA\Schema(
     *          type="integer"
     *       )
     *   ),
     *   @OA\Parameter(
     *        name="name",
     *        description="Product Name",
     *        example="Demo Iphone 14 Pro Max",
     *        in="query",
     *        @OA\Schema(
     *          type="string"
     *        )
     *   ),
     *   @OA\Parameter(
     *        name="price",
     *        in="query",
     *        description="Demo Product Price",
     *        example="1111.99",
     *        @OA\Schema(
     *          type="string"
     *        )
     *   ),
     *   @OA\Parameter(
     *        name="quantity",
     *        description="Product Quantity",
     *        example="30",
     *        in="query",
     *        @OA\Schema(
     *          type="string"
     *        )
     *  ),
     *  @OA\Parameter(
     *        name="product_variants",
     *        description="Product Variants",
     *        in="query",
     *        example="separate ids by comma eg. 1,2,3 etc",
     *        @OA\Schema(
     *          type="string"
     *        )
     *  ),
     * 
     *  @OA\Response(
     *      response=201, 
     *      description="product updated successfully",
     *      @OA\MediaType(
     *         mediaType="application/json",
     *      ),
     *   ),
     *   @OA\Response(
     *      response=401,
     *      description="Unauthorized",
     *   ),
     *   @OA\Response(
     *       response=403,
     *        description="Forbidden"
     *    )
     * ),
     * Update the specified resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @param  \App\Models\Product  $product
     * @return \Illuminate\Http\Response
     */
    public function update(Request $request, Product $product)
    {
        // Validate Request
        $validator = Validator::make($request->only('name', 'price', 'quantity', 'product_variants', 'image'), [
            'name' => 'required|min:3',
            'price' => 'required|numeric|gt:0',
            'quantity' => 'required|numeric|min:0',
            'image' => 'nullable|image|mimes:jpg,png,jpeg|max:2048',
            'product_variants' => 'required|min:1'
        ]);

        // Check if validation failed
        if ($validator->fails()) {
            return $this->sendError($validator->errors(), 'Incomplete Data Provided', 422);
        }

        // Update Product
        $product->update($validator->validated());

        $product_variants = explode(',', rtrim($validator->validated()['product_variants'], ","));

        $product->variants()->sync($product_variants);

        // Check if Image is part of request
        if ($request->hasFile('image')) {
            $mediaItems = $product->getFirstMedia('images');
            if (!$mediaItems == null) {
                $mediaItems->delete(); // delete old image
            }
            $product->addMediaFromRequest('image')->toMediaCollection('images');
        }


        return $this->sendResponse(new ProductResource($product), 'Product Updated Successfully', 201);
    }

    /**
     * @OA\Delete(
     *   path="/api/v1/products/{id}",
     *   operationId="deleteProduct",
     *   summary="Delete Product",
     *   tags={"Products"},
     *   description="Delete a product",
     *   security={ {"BearerAuth": {} }},
     *   @OA\Parameter(
     *       name="id",
     *       description="Product id",
     *       required=true,
     *       in="path",
     *       @OA\Schema(
     *          type="integer"
     *       )
     *   ),
     *   @OA\Response(
     *      response=201, 
     *      description="Product deleted successfully",
     *      @OA\MediaType(
     *         mediaType="application/json",
     *      ),
     *   ),
     *   @OA\Response(
     *      response=401,
     *      description="Unauthorized",
     *   ),
     *   @OA\Response(
     *       response=403,
     *        description="Forbidden"
     *    )
     * ),
     * Remove the specified resource from storage.
     *
     * @param  \App\Models\Product  $product
     * @return \Illuminate\Http\Response
     */
    public function destroy(Product $product)
    {
        $product->delete();

        return $this->sendResponse([], 'Product Deleted Successfully', 201);
    }
}